Implement class inheritance and add mock data for all classes

+++ src/main/java/iss/tim4/domain/dto/driver/document/DriverDocumentDTOResult.java
@@ -13,7 +13,7 @@ public class DriverDocumentDTOResult {
     private Long id;
     private String name;
     private String documentImage;
-    private Long driverId;
+    private Integer driverId;
 
     public DriverDocumentDTOResult(DriverDocument driverDocument) {
         this.id = driverDocument.getId();
